# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: Occount Kiosk Backend API
  description: |
    키오스크 결제 시스템 백엔드 API 명세

    ## 인증 방식
    - **AUTHENTICATED**: `Authorization: Bearer <token>` 헤더 필수
    - **OPTIONAL_AUTH**: 토큰이 있으면 회원으로 처리, 없으면 비회원으로 처리. 잘못된 토큰은 401 반환
    - **PERMIT_ALL**: 인증 불필요
    - **ADMIN_ONLY**: 관리자 권한 토큰 필요

    ## 키오스크 헤더
    - `X-Kiosk-Id`: 주문 생성·취소 시 필수. 키오스크 식별자

    ## 토큰 발급
    - 회원: `POST /api/v3/auth/login`
    - 키오스크 사용자: `POST /api/v3/auth/kiosk/login`
    - 발급된 토큰은 응답 `Authorization` 헤더에 포함됨

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local (API Gateway)
...
paths:
...
  /api/v3/ari-pick:
    get:
      tags: [AriPick]
      operationId: getAripickList
      summary: 아리픽 제안 목록 조회
      description: 제안 목록을 조회합니다. (PERMIT_ALL)
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickListResponse'

    post:
      tags: [AriPick]
      operationId: createAripick
      summary: 아리픽 제안 등록
      description: 로그인 사용자가 식약처 typeNSeq 기반으로 제안을 등록합니다. (AUTHENTICATED)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAripickRequest'
      responses:
        "201":
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickResponse'
        "400":
          description: 매점 물품 수칙 위반
        "503":
          description: 식약처 연동 실패
        "401":
          description: 인증 실패
  /api/v3/ari-pick/foods:
    get:
      tags: [AriPick]
      operationId: searchAripickFoods
      summary: 아리픽 식품 검색
      description: 키워드로 식약처 제품명을 검색합니다. (PERMIT_ALL)
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickFoodSearchResponse'
        "503":
          description: 식약처 연동 실패
  /api/v3/ari-pick/blocked-keywords:
    get:
      tags: [AriPick]
      operationId: getAripickBlockedKeywords
      summary: 아리픽 금지 키워드 목록
      description: 관리자가 등록된 금지 키워드를 조회합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickBlockedKeywordListResponse'
        "403":
          description: 관리자 권한 없음

    post:
      tags: [AriPick]
      operationId: createAripickBlockedKeyword
      summary: 아리픽 금지 키워드 등록
      description: 관리자가 금지 키워드를 등록합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAripickBlockedKeywordRequest'
      responses:
        "201":
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickBlockedKeywordResponse'
        "400":
          description: 금지 키워드가 비어 있음
        "409":
          description: 이미 등록된 금지 키워드
        "403":
          description: 관리자 권한 없음
  /api/v3/ari-pick/blocked-keywords/{keywordId}:
    delete:
      tags: [AriPick]
      operationId: deleteAripickBlockedKeyword
      summary: 아리픽 금지 키워드 삭제
      description: 관리자가 금지 키워드를 삭제합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      parameters:
        - name: keywordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 성공
        "403":
          description: 관리자 권한 없음
  /api/v3/ari-pick/{proposalId}:
    get:
      tags: [AriPick]
      operationId: getAripick
      summary: 아리픽 제안 단일 조회
      description: proposalId로 단일 제안을 조회합니다. (PERMIT_ALL)
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickResponse'
        "404":
          description: 제안을 찾을 수 없음

    delete:
      tags: [AriPick]
      operationId: deleteAripickByUser
      summary: 아리픽 제안 삭제 (작성자)
      description: 로그인 사용자가 본인 제안을 삭제합니다. (AUTHENTICATED)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 성공
        "401":
          description: 인증 실패
        "403":
          description: 본인 제안이 아님
        "404":
          description: 제안을 찾을 수 없음
  /api/v3/ari-pick/{proposalId}/approve:
    patch:
      tags: [AriPick]
      operationId: approveAripick
      summary: 아리픽 제안 승인
      description: 제안을 승인 상태로 변경합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 승인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickResponse'
        "403":
          description: 관리자 권한 없음
        "404":
          description: 제안을 찾을 수 없음
  /api/v3/ari-pick/{proposalId}/reject:
    patch:
      tags: [AriPick]
      operationId: rejectAripick
      summary: 아리픽 제안 거절
      description: 제안을 거절 상태로 변경합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 거절 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickResponse'
        "403":
          description: 관리자 권한 없음
        "404":
          description: 제안을 찾을 수 없음
  /api/v3/ari-pick/{proposalId}/pending:
    patch:
      tags: [AriPick]
      operationId: pendingAripick
      summary: 아리픽 제안 PENDING 변경
      description: 제안을 PENDING 상태로 변경합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickResponse'
        "403":
          description: 관리자 권한 없음
        "404":
          description: 제안을 찾을 수 없음
  /api/v3/ari-pick/{proposalId}/admin:
    delete:
      tags: [AriPick]
      operationId: deleteAripickByAdmin
      summary: 아리픽 제안 삭제 (관리자)
      description: 관리자가 제안을 삭제합니다. (ADMIN_ONLY)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 성공
        "403":
          description: 관리자 권한 없음
        "404":
          description: 제안을 찾을 수 없음
  /api/v3/ari-pick/{proposalId}/like:
    post:
      tags: [AriPick]
      operationId: toggleAripickLike
      summary: 아리픽 좋아요 토글
      description: 좋아요를 토글합니다. (AUTHENTICATED)
      security:
        - BearerAuth: []
      parameters:
        - name: proposalId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 토글 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AripickLikeToggleResponse'
        "401":
          description: 인증 실패
        "404":
          description: 제안을 찾을 수 없음
...
  /api/v3/orders:
    post:
      tags: [Order]
      operationId: createOrder
      summary: 주문 생성
      description: |
        주문을 생성합니다. (OPTIONAL_AUTH)

        - `X-Kiosk-Id` 헤더가 없으면 기본값 `1`을 사용합니다.
        - 토큰이 있으면 회원 주문, 없으면 비회원 주문으로 처리합니다.
        - 요청 즉시 `PROCESSING` 상태로 응답하며, 실제 처리는 비동기로 수행됩니다.
        - 상품 검증, 가격 확정, 재고 차감은 비동기로 처리됩니다.
        - 재고 부족 또는 상품 검증 실패 시 최종 상태는 `FAILED` 입니다.
        - 최종 결과는 `GET /api/v3/orders/{orderId}/stream` (SSE) 로 실시간 수신하거나,
          `GET /api/v3/orders/{orderId}` 를 폴링하여 확인합니다.
      parameters:
        - name: X-Kiosk-Id
          in: header
          required: false
          description: 키오스크 식별자. 없으면 `1`
          schema:
            type: string
            default: "1"
          example: "1"
      security:
        - {}
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
            example:
              items:
                - item_id: 1
                  quantity: 2
      responses:
        "202":
          description: 주문 접수 (PROCESSING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                order_id: 1001
                status: "PROCESSING"
                failure_reason: null
        "400":
          description: 요청 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: 잘못된 토큰
  /api/v3/orders/{orderId}:
    get:
      tags: [Order]
      operationId: getOrder
      summary: 주문 조회
      description: |
        주문 상태를 단건 조회합니다. (PERMIT_ALL)

        - 재고 부족, 상품 미존재 등 재고 단계 실패는 `FAILED` 로 종료됩니다.
        - 최종 상태: `COMPLETED`, `FAILED`, `CANCELLED`, `COMPENSATION_FAILED`, `TIMED_OUT`
        - 실시간 상태 수신이 필요한 경우 `GET /api/v3/orders/{orderId}/stream` (SSE) 를 사용하세요.
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
            format: int64
          example: 1001
      security:
        - {}
      responses:
        "200":
          description: 주문 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                order_id: 1001
                status: "COMPLETED"
                failure_reason: null
        "404":
          description: 주문을 찾을 수 없음
  /api/v3/orders/{orderId}/stream:
    get:
      tags: [Order]
      operationId: streamOrder
      summary: 주문 상태 실시간 스트리밍 (SSE)
      description: |
        Server-Sent Events(SSE) 방식으로 주문 상태 변경을 실시간 수신합니다. (PERMIT_ALL)

        - 연결 즉시 현재 주문 상태를 전송합니다.
        - 각 메시지는 `event: <event_type>` 형식으로 전달됩니다.
        - 이벤트 타입은 `order_accepted`, `payment_requested`, `completed`, `failed`, `cancel_requested`, `cancelled`, `timed_out` 중 하나입니다.
        - 주문이 최종 상태(`completed`, `failed`, `cancelled`, `timed_out`)에 도달하면 서버가 연결을 종료합니다.
        - 주문 생성 직후 SSE 연결을 맺어 비동기 처리 결과를 수신하는 용도로 사용합니다.
      parameters:
        - name: orderId
          in: path
          required: true
          description: 주문 ID
          schema:
            type: integer
            format: int64
          example: 1001
      security:
        - {}
      responses:
        "200":
          description: SSE 스트림 연결 성공
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE 메시지는 아래 형식으로 전달됩니다.

                  - `event`: 이벤트 타입
                  - `data`: 이벤트 데이터 (JSON). 정상 이벤트는 `{}`, 실패 이벤트는 `{"failure_reason": "..."}` 포함
              examples:
                accepted:
                  summary: 주문 접수 직후
                  value: |
                    event: order_accepted
                    data: {}

                paymentRequested:
                  summary: 결제 요청 단계
                  value: |
                    event: payment_requested
                    data: {}

                completed:
                  summary: 주문 완료
                  value: |
                    event: completed
                    data: {}

                failed:
                  summary: 주문 실패
                  value: |
                    event: failed
                    data: {"failure_reason": "재고 부족"}

                cancelled:
                  summary: 주문 취소 완료
                  value: |
                    event: cancelled
                    data: {}
        "404":
          description: 주문을 찾을 수 없음
  /api/v3/orders/{orderId}/cancel:
    post:
      tags: [Order]
      operationId: cancelOrder
      summary: 주문 취소
      description: |
        주문을 취소합니다. (OPTIONAL_AUTH)

        - `X-Kiosk-Id` 헤더가 없으면 기본값 `1`을 사용합니다.
        - 주문을 생성한 키오스크와 동일한 `X-Kiosk-Id` 여야 취소가 허용됩니다.
        - 회원/비회원 구분 없이 키오스크 ID만으로 소유권을 검증합니다.
      parameters:
        - name: orderId
          in: path
          required: true
          description: 취소할 주문 ID
          schema:
            type: integer
            format: int64
          example: 1001
        - name: X-Kiosk-Id
          in: header
          required: false
          description: 키오스크 식별자. 없으면 `1`
          schema:
            type: string
            default: "1"
          example: "1"
      security:
        - {}
        - BearerAuth: []
      responses:
        "200":
          description: 취소 요청 접수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                order_id: 1001
                status: "CANCEL_REQUESTED"
                failure_reason: "사용자에 의해 주문이 취소되었습니다"
        "401":
          description: 잘못된 토큰
        "403":
          description: 다른 키오스크에서 생성된 주문
        "404":
          description: 주문을 찾을 수 없음
        "409":
          description: 취소 불가 상태의 주문

  # ───────────────────────────────────────────────
  # Wallet
  # ───────────────────────────────────────────────
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <JWT 토큰>"
  schemas:
...
    CreateAripickRequest:
      type: object
      required: [type_n_seq, reason]
      properties:
        type_n_seq:
          type: integer
          format: int64
        reason:
          type: string
    AripickFoodSearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AripickFoodSearchItemResponse'
    AripickFoodSearchItemResponse:
      type: object
      properties:
        type_n_seq:
          type: integer
          format: int64
        name:
          type: string
        company:
          type: string
        kcal_info:
          type: string
    CreateAripickBlockedKeywordRequest:
      type: object
      required: [keyword]
      properties:
        keyword:
          type: string
    AripickBlockedKeywordResponse:
      type: object
      properties:
        keyword_id:
          type: integer
          format: int64
        keyword:
          type: string
        registered_date:
          type: string
          format: date
    AripickBlockedKeywordListResponse:
      type: object
      properties:
        keywords:
          type: array
          items:
            $ref: '#/components/schemas/AripickBlockedKeywordResponse'
    AripickListResponse:
      type: object
      properties:
        aripick_items:
          type: array
          items:
            $ref: '#/components/schemas/AripickResponse'
    AripickResponse:
      type: object
      properties:
        proposal_id:
          type: integer
          format: int64
        name:
          type: string
        reason:
          type: string
        proposer_id:
          type: integer
          format: int64
        proposal_date:
          type: string
          format: date
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        like_count:
          type: integer
    AripickLikeToggleResponse:
      type: object
      properties:
        proposal_id:
          type: integer
          format: int64
        liked:
          type: boolean
        like_count:
          type: integer
...
    OrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: 주문 상품 목록 (`orderInfos` 별칭 허용)
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemRequest:
      type: object
      required: [item_id, quantity]
      properties:
        item_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    OrderResponse:
      type: object
      properties:
        order_id:
          type: integer
          format: int64
          description: 주문 ID
        status:
          $ref: '#/components/schemas/OrderStatus'
        failure_reason:
          type: string
          nullable: true
          description: 실패/취소 사유
    OrderStatus:
      type: string
      description: |
        주문 상태 값.

        **최종 상태** (SSE 연결이 자동 종료됨):
        - `COMPLETED` — 주문 완료
        - `FAILED` — 주문 실패 (재고 부족, 상품 미존재 등)
        - `CANCELLED` — 주문 취소 완료
        - `COMPENSATION_FAILED` — 보상 트랜잭션 실패
        - `TIMED_OUT` — 처리 시간 초과

        **중간 상태**:
        - `PENDING` — 초기 상태
        - `PROCESSING` — 처리 중
        - `CANCEL_REQUESTED` — 취소 요청됨
        - `COMPENSATING` — 보상 처리 중
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        - CANCEL_REQUESTED
        - COMPENSATING
        - CANCELLED
        - COMPENSATION_FAILED
        - TIMED_OUT

    # ── Wallet ─────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: 오류 메시지